L14HW22 MusicStyles package refactoring
