+++ src/main/java/ua/vodafone/lecture14/musicstyles/ClassicMusic.java
@@ -1,4 +1,4 @@
-package ua.vodafone.lecture14;
+package ua.vodafone.lecture14.musicstyles;
 
 public class ClassicMusic extends MusicStyles {
     private final String musician;
