+++ src/main/java/ua/vodafone/lecture14/musicstyles/MusicRunner.java
@@ -1,4 +1,4 @@
-package ua.vodafone.lecture14;
+package ua.vodafone.lecture14.musicstyles;
 
 public class MusicRunner {
     public static void main(String[] args) {
