+++ src/main/java/ua/vodafone/lecture14/musicstyles/MusicStyles.java
@@ -1,4 +1,4 @@
-package ua.vodafone.lecture14;
+package ua.vodafone.lecture14.musicstyles;
 
 public abstract class MusicStyles {
     public abstract void playMusic();
